requirements.txt
playwright==1.41.2
pytest-playwright==0.4.4
python-dotenv==1.0.1
six==1.16.0
urllib3==1.26.18